+++ common/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "xmtp_common"
-edition = "2021"
+edition = "2024"
 version.workspace = true
 license.workspace = true
 
@@ -8,14 +8,14 @@ license.workspace = true
 arc-swap.workspace = true
 async-trait.workspace = true
 bytes.workspace = true
-futures.workspace = true
+futures = { workspace = true, features = ["std"] }
 hex.workspace = true
 rand.workspace = true
+rand_chacha.workspace = true
 thiserror.workspace = true
-tokio = { workspace = true, features = ["time"] }
+tokio = { workspace = true, features = ["time", "rt"] }
 tracing.workspace = true
 web-time.workspace = true
-xmtp_cryptography.workspace = true
 
 # optional/test-utils/dev-deps
 once_cell = { workspace = true, optional = true }
