+++ result
@@ -1,0 +1,1 @@
+/nix/store/yzsg1gp3pn5y7m9kh7pk9g79sb1q91hq-bindings_wasm-0.0.1
