+++ xmtp_api_d14n/src/lib.rs
@@ -37,6 +37,7 @@ pub mod tests {
 
     pub type TestV3Client = V3Client<TestClient>;
     pub type TestD14nClient = D14nClient<TestClient, TestClient>;
+
     // Execute once before any tests are run
     #[cfg_attr(not(target_arch = "wasm32"), ctor::ctor)]
     #[cfg(not(target_arch = "wasm32"))]
