+++ xmtp_mls/Cargo.toml
@@ -38,8 +38,9 @@ bench = [
   "xmtp_common/bench",
 ]
 default = ["grpc-api"]
-grpc-api = ["dep:xmtp_api_grpc"]
-http-api = ["dep:xmtp_api_http"]
+grpc-api = ["dep:xmtp_api_grpc", "xmtp_api_d14n?/grpc-api"]
+http-api = ["dep:xmtp_api_http", "xmtp_api_d14n?/http-api"]
+d14n = ["dep:xmtp_api_d14n"]
 update-schema = ["toml"]
 
 [dependencies]
@@ -84,6 +85,7 @@ const_format = { workspace = true, optional = true }
 ethers = { workspace = true, features = ["openssl"], optional = true }
 fdlimit = { workspace = true, optional = true }
 toml = { version = "0.8.4", optional = true }
+xmtp_api_d14n = { workspace = true, optional = true }
 xmtp_api_http = { path = "../xmtp_api_http", optional = true }
 
 # Test/Bench Utils
@@ -138,6 +140,8 @@ fdlimit = { workspace = true }
 once_cell.workspace = true
 rstest = { workspace = true, features = ["async-timeout"] }
 xmtp_db = { workspace = true, features = ["test-utils"] }
+xmtp_api_d14n = { workspace = true, features = ["test-utils"] }
+
 
 [target.'cfg(not(target_arch = "wasm32"))'.dev-dependencies]
 ctor.workspace = true
